Added multiple logouts and distinct profiles

+++ src/main/java/org/example/Models/CheckingAccount.java
@@ -1,18 +1,20 @@
 package org.example.Models;
 
+import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
+import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleIntegerProperty;
 
 public class CheckingAccount extends Account{
     // the number of transactions a client is allowed to do per day
-    private final IntegerProperty transactionLimit;
+    private final DoubleProperty transactionLimit;
 
     public CheckingAccount(String owner, String accountNumber, double balance, int tLimit) {
         super(owner, accountNumber, balance);
-        this.transactionLimit = new SimpleIntegerProperty(this, "Transaction Limit", tLimit);
+        this.transactionLimit = new SimpleDoubleProperty(this, "Transaction Limit", tLimit);
     }
 
-    public IntegerProperty transactionLimitProperty() {
+    public DoubleProperty transactionLimitProperty() {
         return this.transactionLimit;
     }
 }
